blockify/deblockify: Make data-manipulation routines async safe
This is/was a source of a nasty bug. Everything worked so long as
hosts weren't scheduled for an immediate re-check (i.e. returned non-OK
but was previously OK) and the result came in through the module.

What happened was that the scheduling routine picked up the scheduled
check while the memory it was pointing to (residing on the heap and
thus shared between threads) was being modified by the module. This
didn't show up in the early tests, because the stupid plugin being
used in those always returned OK.

Conceptually this patch is much larger than the diffstat output shows,
so possibly we'll run into other problems with this. Lots and lots of
further testing is required.

Signed-off-by: Andreas Ericsson <ae@op5.se>