handle_service_result(): Handle malformed packets and missing perfdata
Previously the module would cause Nagios to crash when it either
couldn't find the service it was looking for or if the pointer to
the perfdata string was set to a value but overshot the buffer-space.

This is a temporary fix though. In the (near, I hope) future
the deblockify() function in data.c should be used instead.

Signed-off-by: Andreas Ericsson <ae@op5.se>