Don't increment the event counter for non-existing objects.
If something goes seriously wrong with configuration propagation
we could have hosts running configurations with hosts and/or services
that no longer exist on the noc side. When we receive a check-event
for such an object we musn't increment the event counter, since it's
not, in fact, a valid event for us.

This also mitigates one possible DoS vector, where a malicious user
could send a continuous stream of invalid invalid results that would
cause the noc node to never exit the event-receiving loop and thrash
disks at a very high pace.

Signed-off-by: Andreas Ericsson <ae@op5.se>