push_logs.sh: A poor mans log sync
push_logs.sh will try its very best to figure out the
bare minimum of logs to push to a central master server,
perhaps after a serious outage or some such. It's intended
to be used to send the minimum of logs required to
generate reports with and committed to the merlin project
in the hopes that it will be useful to someone.

Signed-off-by: Andreas Ericsson <ae@op5.se>