Grok logging variables too
We had a function called "log_grok_var()" which was
designed for the logging api to get its settings from,
but it never got called so it was pretty much worthless.
This patch fixes it so that module and daemon alike
both can have a proper log_file (and log_levels)
setting in their respective compounds.

Signed-off-by: Andreas Ericsson <ae@op5.se>