Plug memory leak in net_input() from networking api
We used to blindly allocate as much memory as the header
proclaimed the packet required and we didn't free() it
if we took the error-path and returned early.

This patch makes use of a static stack-allocated buffer
instead, and also makes sure we never try to write beyond
the end of it by checking the size we're about to try to
read.

It also makes us bail early in case we've got a bad
packet header (fe, if the protocol version is one we know
nothing about).

Signed-off-by: Andreas Ericsson <ae@op5.se>