Make events read from ipc socket re-usable
Previously we used to let the sender (send_ipc_data) complete
the packet to send, which wasn't necessarily the best approach
to take as it prevented us from plugging in additional
functionality without fiddling with the transport layer.

This makes for a cleaner approach. Completing the packet is
now left to "handle_ipc_data()", which reads the body of the
packet into a dynamic buffer and concatenates the two.

This means that additional calls can be made for the data
we just read without incurring undue overhead in any such
new functionality.

Signed-off-by: Andreas Ericsson <ae@op5.se>