Properly activate/deactivate nodes on takeover
This change has been sitting in the tree for quite some time,
so I'm no longer 100% sure why it was made. The patch *looks*
sane though, so I'm going ahead and committing it on the (very
vague) basis that I probably knew what I was doing when I made
the change.

Signed-off-by: Andreas Ericsson <ae@op5.se>