ipc: Avoid segfault when no ipc socket is configured
We used to want to take the strlen() of the path to
the ipc socket without first checking if it was NULL,
which would cause us to crash when no such socket was
configured.

This patch corrects that, adding a check and logging
a warning for when no socket path has been configured.

Signed-off-by: Andreas Ericsson <ae@op5.se>