ipc: Add ipc_read_event()
It does exactly what it sounds like, and should replace
ipc_reap() later (although the latter is left behind for
now). In addition, this makes the need for checks for
connectedness in ipc_is_connected() redundant, as the
way to determine whether or not a unix domain socket is
connected is to find out if it has 0 bytes (as returned
by recv()/read() et al) of data available, but poll()
and/or select() and their ilk claims it has inbound
data waiting to be read.

Signed-off-by: Andreas Ericsson <ae@op5.se>