Revamp socket polling and use packet-based functionality in daemon
Previously, network sockets and the ipc socket weren't polled
simultaneously, leading to possible congestion in either kind
while the other was being watched. With this patch, the polling
system is changed so that all possible interesting sockets are
watched simultaneously. This means that when nothing's happening,
we do nothing (rather than hop back and forth between the two
different kinds of sockets).

While we're at it, we convert much of the remaining system to
use a packet-based mode of operation, since that's the most
natural state of being for us anyway (everything we have to
do always involves the packet body and at least a subset of
the protocol header).

This patch is larger than what I'd like, but the two changes
evolved simultaneously and one happened to solve the other,
and the effort required to split the two would be larger than
any possible gain that could be had from it, especially
considering it didn't work very well before this patch anyway.

Signed-off-by: Andreas Ericsson <ae@op5.se>