ipc.c: Wait a short time before dismissing the ipc socket as unwritable
With 0 milliseconds wait-time before discarding packets, we received
42578 / 1000000 packets in 7.87 seconds time. With 100ms wait time
on polling for writability, we had 1000000 / 1000000 in 3m 10.044s
which is obviously an improvement. A 100 millisecond delay is not
a big deal, and precision is more important.

Currently we can accept 5262 events per second in the daemon, while
at the same time writing them to database. In terms of Nagios
installation size, where we calculate an average check interval of
5 minutes, that yields 1578600 monitored items (services and hosts).
A little less with allowed overhead for ancillary data, but for a
rough estimate, it will suffice.

Signed-off-by: Andreas Ericsson <ae@op5.se>