Make ipc_sock a static variable
Since we added some logic to return the correct socket
depending on whether or not it's connected, we had better
make sure that logic is always used.

Since the only offender was one function in daemon.c, we
convert it to use the already coded wrapper in ipc.c instead,
and thus this change is really quite small.

As an added bonus, this also shuts up one sparse warning.

Signed-off-by: Andreas Ericsson <ae@op5.se>