daemon.c: Fix two sparse warnings
io_poll_sockets() should never be called from outside daemon.c
and it wasn't declared, thus causing a warning. This makes it
static.

We used to allocate memory for a packet structure, but then
we shadowed it anyway with a stack-allocated structure in a
narrower scope. This patch removes the heap-allocated packet
structure pointer and its allocation code, removing a sparse
warning about the shadowed variable.

Signed-off-by: Andreas Ericsson <ae@op5.se>