sql.c: Introduce sql_quote(), wrapping dbi_conn_quote_string_copy()
libdbi-0.8.1, which is available on our customer's systems, does
not have dbi_conn_escape_string_copy(), which was introduced in
libdbi-0.8.2. Since libdbi lacks a proper version macro we can't
determine the version at compile-time, and we have to fall back
to wrapping dbi_conn_quote_string_copy() and remove the two
quotation marks for the sql_escape() function.

Signed-off-by: Andreas Ericsson <ae@op5.se>