hooks.c: Add a generic hook
Since the only reason we're using special hooks for host and
service check results is to make sure we never transmit
unfinished check-results (as the hook gets called twice; Once
before the check runs and once after it's done running), we
can make the general hook a lot simpler and let it simply
transfer everything it catches.

We'll probably run into a few segfaults with this approach,
since I'm sure there are other broker events which have the
same call-twice-run-once philosophy as check-results, but
we'll catch those when we get there (and writing hooks for
that will be super-simple, which is nice).

Signed-off-by: Andreas Ericsson <ae@op5.se>