Take control packets into account in packet_size() macros
Otherwise using packet_size() requires that callers are aware
of which packets are control packets, which goes against the
idea that writers should have a simple means of determining
how much data to send by using this rather simple macro.

We also add typedef's for merlin_event and merlin_header.

Signed-off-by: Andreas Ericsson <ae@op5.se>