Revamp ipc socket dance for daemon
This lets the daemon keep track of both the listening and the
connected socket for the ipc socket code, which helps quite a
lot when debugging intermittent ipc errors (well, they seemed
intermittent anyways).

Since this code seems to work reliably, we use it and let it
go at that.

Signed-off-by: Andreas Ericsson <ae@op5.se>