Honor configured database name in all queries
Previously we used the hardcoded monitor_gui, but since we
allow the database name to be configured we should also honor
it. This patch does just that, by adding a helper function to
sql.[ch] to return the configured database name (or "merlin"
if no name is configured), and updating the queries to take
that configured name into account.

While at it, we also fix the prototype for sql_table_name().

Signed-off-by: Andreas Ericsson <ae@op5.se>