Revamp daemonize interface
We don't need to use logging there, since daemonizing always
happens before we've released stderr and stdout. This way,
we can also make it a free-standing helper interface.

We also take care of some splint and sparse warnings while
we're at it, and rework a few of the functions slightly.

Signed-off-by: Andreas Ericsson <ae@op5.se>