install script: Check for files being present when we need them
We used to check for the necessary files to install regardless of
which parts the user asked us to install, but since the install
script is runnable from the installed-to directory (which doesn't
necessarily have all the required files), this didn't work very
well.

With this patch, the install-script only checks for required
files when it actually needs them, and so they don't get in the
way of users wanting to do just the nagios-configuration modifying
parts, or re-installing the database after an installation is
completed.

Signed-off-by: Andreas Ericsson <ae@op5.se>