module: Properly handle the 'module' config compound
Previously we used to check 'nested' instead of 'vars', which
was fairly stupid, as it'd mean we'd try to iterate over as
many variables as the compound has subcompounds, meaning we'd
miss logfile settings and other goodies.

This patch fixes that, and also moves the "module loaded"
message until after the configuration has been read, so that
the message ends up in the logfile instead of on stdout, which
in most normal cases will be close()d.

Signed-off-by: Andreas Ericsson <ae@op5.se>