object_importer: Keep host_names as strings in service slave objects
Since we still link the 'pure' slave objects to their masters by id,
we let the object->host links remain in their textual representation
in the service slave objects. This saves us two hash-lookups as we
no longer need to find the host id when parsing the object the first
time, and another one since we don't need to convert the id to a
name later when post-mangling service slave objects.

Signed-off-by: Andreas Ericsson <ae@op5.se>