daemon: Bail out noisily when sql initialization fails
sql_init() is very thorough in returning a proper fail/success
return code, but that return code was never used. When the
SQL connection did indeed fail, or some other aspects of the
database initialization failed to work properly, we would
previously segfault when trying to use the (nonexistant)
database connection.

This patch fixes that by aborting on any SQL initialization
error.

Reported-by: Mario Eckl <mario.eckl@gmail.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>