object_importer: Retain status data and object id's across imports
Previously we used to truncate all object tables and re-do the
entire import from scratch. Since that means discarding all the
state data it's not such a stellar idea. This patch provides a
means for pre-loading the index tables and re-using the same
object id's when re-running an import. The query is modified
to use "UPDATE ... SET key = value, key2 = value2" etc instead
of the previous "REPLACE INTO" query.

The downsides of this patch is that old objects will be retained
in the database indefinitely. Fixing that will be a separate
patch.

Signed-off-by: Andreas Ericsson <ae@op5.se>