object_importer: Purge old objects from database
Since the previous code retained old objects indefinitely, we could
well end up with a huge database that would, eventually, cause us
to bail out due to the sheer load of the number of objects in it.

This patch causes objects in the two protected tables that weren't
imported by this particular run to be purged when we're done
writing out the new objects.

Signed-off-by: Andreas Ericsson <ae@op5.se>