object_importer: Learn how to import status.sav files
Since it's quite boring to wait for status updates to happen
when one first starts using Merlin + Ninja to monitor things,
we can adapt the object importer slightly to make it capable
of importing status.sav files as well. This also has the
added benefit that old comments will get automatically
preserved, which is nice.

Signed-off-by: Andreas Ericsson <ae@op5.se>