sql: Add sql_get_result()
This is just a stupid wrapper to return the stored dbi_result
handle from the db struct we're using in sql.c. This is necessary
to make it possible for us to issue queries where we actually
care about the result, which in turn is required for us to be
able to prime the merlin daemons internal status tables with
retention data so we can know when the last statechange
occurred, which in turn is required for the GUI to know how
long a particular host or service has been in a particular state.

Lots of turns there...

Signed-off-by: Andreas Ericsson <ae@op5.se>