Add dbitest.c, a sample implementation for reading status data
dbitest.c is just a really stupid program that reads all the
status data from the database into memory as a sort of retention
data fallback. It's entirely possible that we could do without
it in most cases, as it *should* be reasonably quick for us to
hit the database to find the status variables we need. However,
if the database is on a remote host, each query will invoke a
small but significant latency, and we'd rather avoid that.

Signed-off-by: Andreas Ericsson <ae@op5.se>