Introduce 'status.c'
This is a (slight) rework of the dbitest.c program. Its intended
use is to preload status data from a previously populated
database so that the Merlin daemon can tell whether a state is
new or not.

In order to do that when Nagios has been reloaded, we need to
add support for running the import from Merlin, so we know
when it's time to actually prime the status tables.

Signed-off-by: Andreas Ericsson <ae@op5.se>