protocol.h: Introduce CTRL_PATHS
This particular control code is used when the module wants to send
paths to certain files to the daemon. We do it this way since Nagios
implements default paths for very nearly all the files it uses, and
we need the correct paths to the objects.cache and status.log files
in order to prime the database with that information, as well as in
order to send configuration to poller nodes (if we're ever supposed
to do that in the future).

Signed-off-by: Andreas Ericsson <ae@op5.se>