Implement auto-import of objects and status when module connects
Actually, we auto-import objects whenever the module sends us the
path information, but since it does that (or at least tries to)
when it connects, it amounts to the same thing.

This patch adds both the module and daemon support for the new
CTRL_PATHS control code, thus allowing the module to send the
full paths to objects.cache, nagios.cfg and status.log to the
daemon, which then runs the appropriate command to populate the
database. That is, assuming "the appropriate command" is actually
a PHP-script sitting on my laptop. Configuration support for the
import command is as yet missing.

Signed-off-by: Andreas Ericsson <ae@op5.se>