ipc.c: Don't be quite so aggressive when trying to read ipc events
If we *know* reading will fail, we give up early. Some other checks
are also added which means we never try to write to a socket that
isn't poll()'able, and we return a proper exit code in case polling
didn't work, which means we shouldn't set up and tear down the ipc
socket connection quite so often anymore.

Signed-off-by: Andreas Ericsson <ae@op5.se>