Deprecate the 'mode' configuration variable
Since we can automatically deduce what we're supposed to do based
on the configured peers and pollers we have, we really don't need
this extra hint to tell us what we are.

This patch also adds some infrastructure for deprecating config
parameters in the future, with an expiration date to print when
warning people that they're using (soon) unsupported configuration
options.

Signed-off-by: Andreas Ericsson <ae@op5.se>