daemon.c: Add a handler for network input
This gets called whenever we receive an event from a network node
(well, nearly anyways, we skip pulse packets since the protocol
layer takes care of those itself to mark which nodes are alive
etc).

It's currently awesomely stupid, ignoring errors from the
functions it calls, and all it does is pass the event off to the
database update function and then write it to the ipc socket.

For now though, this suffices, since it proves the point that
the implementation is relatively robust.

Signed-off-by: Andreas Ericsson <ae@op5.se>