daemon.c: Take the path to the import program as a config parameter
Previously this was hardcoded in a way that made it only work on
my laptop, which obviously isn't very userfriendly. Using a config
parameter instead makes much more sense. The default still points
to the old value (my laptop), but with some minor modifications
to the example config this will work splendidly.

Signed-off-by: Andreas Ericsson <ae@op5.se>