status.[ch]: Add public API's to obtain the object state struct
We want the db_updater function to take care of setting the new
status for a host when we receive a check-result, so it's far
better to have it use the already returned pointer to update
the state than it is to do a second binary search with a small
API to update the status, so we publish the object_state struct
as well with this patch.

Since bsearch(3) makes no guarantees about which order it will
pass the arguments, we implement our own binary search to fill
our needs.

Signed-off-by: Andreas Ericsson <ae@op5.se>