status.h: Add some handy macros to handle state and state types
Since 'state' is stored as a single variable in memory, we need
to do some binary operations on it to extract the state and the
type from that variable. It's also handy to have a macro to
concatenate the state, so that it's done the same everywhere
we do it. This patch introduces those macros.

Note that it doesn't actually make sense to store the variables
separate, since we'll need to compare the concatenated state
once we have the reports-module merged into Merlin anyway, and
this way will be slightly faster since the integer we're working
with can be loaded into a single register, as opposed to hitting
memory twice.

Signed-off-by: Andreas Ericsson <ae@op5.se>