Print names of callbacks we don't yet handle in the module
It's thoroughly annoying to only get the id number of the callback
and then having to go look it up for oneself. This patch adds a
small helper function that looks up the name of the callback (or
at least the nagios macro name) in an array and returns that when
given a callback-id within the valid range.

Signed-off-by: Andreas Ericsson <ae@op5.se>