Get rid of the 'is_noc' variable
Since a node can be noc, peer and poller all at the same time, it
really doesn't make sense to try to fit all scenarios into the a
single flag. The logic is now replaced with other, more sensible
options such as checking whether an inactive node was a poller
before trying to take over checks from that poller, or seeing if
we have any pollers before we try to look one up.

Signed-off-by: Andreas Ericsson <ae@op5.se>