control.c: Avoid off-by-one error when disabling checks
The old-time favourite "i = 0; i <= num_entries; i++" where
num_entries is the amount of entries in the array, but the
array is indexed from zero, leading to an out-of-bounds
access on some (but not all) platforms in some cases.

This *might* fix the random crash issues when reloading
Nagios with the Merlin module loaded, and it might also
fix the random crashes we sometimes see on 64-bit archs.

Signed-off-by: Andreas Ericsson <ae@op5.se>