Log version once initialization is completed
We do this by grabbing the version number out of git when
possible, so that bug reports from random people can get
submitted against an exact merlin version.

To this end, we introduce gen-version.sh, which grabs the
git describe output and writes it as a const var to version.c,
which is linked into both the daemon and the module.

Signed-off-by: Andreas Ericsson <ae@op5.se>