db_updater: Fix downtime start/stop events for hosts
Previously the ' AND service_description = "%s"' clause was added
to the query irregardless of whether it was a service downtime or
a host downtime, but the host table has no service_description
field, so the condition made the query impossible and libdbi would
always report an error for host downtime start and stop events.

This patch makes two separate queries of it, which should fix the
issue nicely.

Reported-by: Greg J. Frater <GJFRATER@bechtel.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>