db_updater.c: Make room for more than one check result handler
Previously the stored object state was updated inside the database
update function for each specific state. This meant we couldn't
(easily) run more state-dependent queries later.

This patch moves the object state loading to outside the host and
service check result handlers and adds a comment to the case
table for where someone wanting to extend the check result handling
in a state-dependent way.

Signed-off-by: Andreas Ericsson <ae@op5.se>