Be more careful about printing invalid selections
For some reason it seems we sometimes crash while trying to
disable checks for a selection that doesn't exist. This patch
makes the merlin module bail out early (by returning, not
exiting) when such an invalid selection is encountered. This
will mean a noc's checks may not be properly disabled once its
poller connects, but that's a much smaller problem than having
the entire Nagios instance crashing.

Signed-off-by: Andreas Ericsson <ae@op5.se>