Only execute 'create database' if db do not exist
Using 'create database if not exists' to make sure we do not try to
create a database that allready exist. Also (sorry) check the
db_version to decide if db.sql file should be applied. Currently the
db.sql file is (re-)applied if version do not exist or is equal to 1.

Signed-off-by: Peter Ostlin <peter.ostlin@op5.com>