Lay the groundwork for host/service status events
Previously we've been relying on check result events only,
but it's becoming increasingly clear that it's just not
enough, and that we'll need to support status events as
well. This (large but straight-forward) patch lays the
very basic groundwork for doing just that, by adding a
monitored_object_state structure which we use as an overlay
for the host and service status objects. Since those data
events don't have any actual structs of their own (which
would be damn useful), we're stuck having to implement
our own.

Signed-off-by: Andreas Ericsson <ae@op5.se>