module.c: Send external command events
The groundwork has been in place a long time. There's still
nothing in the other end which understands the event-type,
but this will be absolutely crucial for distributed setups.

Most likely, we'll want to maintain an internal object state
to keep track of which checks are disabled manually and which
are disabled by the distributed checking machinery.

Signed-off-by: Andreas Ericsson <ae@op5.se>