object_importer: Properly support pending states
We mark pending with the special state-code 6 in the merlin
database, which can be handled quite nicely with ninja to
make logical-AND based filters, but Nagios internally marks
it as two separate variables.

This patch checks the state of the two interesting variables
and updates the current_state if either of them is set.

This is part of the fix needed for issue #2520.

Signed-off-by: Andreas Ericsson <ae@op5.se>