Clean up header inclusion
We make shared.h, daemon.h and module.h handle the bulk of all
the header inclusion, saving us a lot of headache in the various
C source files.

The only C sources not including either of the aforementioned
headers are the free-standing API's which do not need to know
about anything merlin-related.

Signed-off-by: Andreas Ericsson <ae@op5.se>