Enable sending of host and service status data events
This patch revamps the hook logic slightly to make logging a whole
lot easier. There's now a single entry-point for all the hooks,
from which we can log a message about which type of callback we
received (good for debugging), and which also makes sure we can
trace possible crashes.

While we're doing that, enabling sending of the host and service
status events is quite simple, so we do just that.

Signed-off-by: Andreas Ericsson <ae@op5.se>