Update config API to match that of reports-module
The linker seems to become utterly confused about having
two symbols with the same name but different signature in
its stack. During certain conditions this leads to a nasty
case of stack corruption. Since the modules are loaded
from userspace allocated memory pages, this almost always
results in a return-into-module stack smashing bug. This
warps the stack frames and makes backtracing nearly
impossible. What little data can be gleaned from the
stack trace is, almost invariably, 100% wrong and very
misleading.

Currently, the only known way to reliably reproduce this
error is to install a package that restarts Nagios via the
init script in its %post scriptlet section. Since that's
the normal mode of restart for blackbox like systems, this
was very bad indeed.

This patch appears to fix it temporarily. A more permanent
fix would be to prefix all public functions with 'merlin'
or some unique string unlikely to be reused by other NEB-
module writers.

Signed-off-by: Andreas Ericsson <ae@op5.se>