module: Plug a leak in {host,service}_status hooks
We allocate one merlin_{host,service}_status object for each host- or
service status update, but the object was never released. Since host
and service status updates are quite frequent, and the object leaked
is quite large, this could quickly mount up to quite a serious issue.

Also plug a small leak in module.c:setup_hash_tables() where we used
to leak the entry-tracker counter. This leak was never serious, as
it only happened on module instantiation, but it's good practice to
avoid even small leaks since it prevents ignorable positives when
using tools such as cppcheck, which was used to find these two.

Signed-off-by: Andreas Ericsson <ae@op5.se>