module: Always send paths when establishing a new ipc connection
Previously we would only send the paths required to run the import
program if we managed to connect to the daemon before any events
were received, and if the daemon was started after the module, no
import would be run.

This fixes it by using the connect/disconnect hooks recently added
to the ipc layer to trigger sending of the paths.

Signed-off-by: Andreas Ericsson <ae@op5.se>