daemon: Bail out noisily if using database without import program
The import program is vital to getting the database to work properly.
Previously we used to fall back to a hardcoded default, which pointed
to a path on my own laptop. Few (if any) users will have their import
program in that location, so it's better if we warn loudly about it
and just exit early.

This fixes issue #2634.

Signed-off-by: Andreas Ericsson <ae@op5.se>