Make recursive inclusion protection macros ANSI C conformant
Previously some of them were named _FILENAMEINCAPS_H_, which
is, technically, a reserved ANSI C keyword since it begins with
an underscore. This changes it to make them use the protection
model used elsewhere in the newer headers instead, namely
INCLUDE_filename_h__, which is guaranteed not to be reserved due
to the two trailing underscores.

Signed-off-by: Andreas Ericsson <ae@op5.se>