daemon: Reap ipc events in bulk
Previously we used to reap one ipc event at a time and then go
back to the polling loop, setting up select fd-sets and all that
jazz. As a result, the overhead of reading a single event from
the ipc socket stacked up quite high.

With this patch, we hope to lower that overhead significantly in
order to not drop events from the module anymore.

This may fix issue #2676.

Signed-off-by: Andreas Ericsson <ae@op5.se>