read() all ipc events in a single operation
Since we need to empty the ipc socket buffer fast, we may as well
special-case the reading and then hop around in a memory-based
buffer to grab the events as we get them. This should ensure
there's always room in the socket buffer for new events to be
inserted.

We also log the time it takes to reap and handle all the events,
which can provide highly beneficial when it comes to identifying
slow queries and/or missing indices in the database.

Signed-off-by: Andreas Ericsson <ae@op5.se>