Handle recurring reads gracefully
Since we use a static function-scoped variable in the ipc reaper
function to indicate the offset into the databuffer, we must make
sure to reset it prior to reading the buffer the second time, and
to make sure we set it to something sensible in case we encounter
an unfinished event in the socket queue.

This patch does just that, hopefully preventing segfaults in the
merlin daemon.

Signed-off-by: Andreas Ericsson <ae@op5.se>