db schema: Remove unused indices
This patch removes 10 indices from the service table, and 9 from the
host table. These indices all had to be updated constantly for each
and every host- and service check and status event. Some events forced
two updates of all the indices.

This slowed down the query processing speed of the sql engine quite
enormously, so we're better off without them.

Signed-off-by: Andreas Ericsson <ae@op5.se>