db_updater: Make processing check results obsolete
The host- and service status events hold very nearly all the info
that the host and service check events contain. By updating output,
long_output and perf_data from those status events, we cut the
amount of queries we need to execute by nearly 60%. This is because
the check result events cause two queries to be run instead of just
one, which ofcourse provides an enormous saving when the database
is busy.

Signed-off-by: Andreas Ericsson <ae@op5.se>