Force-trigger an import every 15 seconds when events go missing
Currently, the module is capable of writing faster than the
daemon is capable of reading. This causes a certain congestion
in the protocol. Until we can solve that in a proper manner,
we fall back on triggering an import of status data and the
object cache whenever we encounter an event that couldn't be
written. This means the database should never lag behind more
than 15 seconds, plus however long the database import script
takes to run. At an absolute maximum, a 20 second delay is
the worst we *should* encounter. Given the latest developments
with the fewer sql queries we're running, this shouldn't happen
very often, but now we'll at least be handling it properly
when it does.

Signed-off-by: Andreas Ericsson <ae@op5.se>