hooks.c: Add the previously missing 'break' statements
When updating the hook coordinator to no longer return the results
of the specific event handlers immediately, the "break" statements
were inadvertantly left out. This caused Nagios to segfault almost
immediately upon calling any eventhandler, which is obviously a
pretty bad thing.

This patch fixes it, adding the missing "break" statements so that
the cases no longer fall through.

Signed-off-by: Andreas Ericsson <ae@op5.se>