importer: Let caller decide when to start and end importing
This is accomplished by adding two new public functions to
object_importer.inc.php; prepare_import() and finalize_import()
which are meant to be called before and after, respectively,
the import actually starts. This will make the import slightly
faster and may also help us avoid the pitfall of removing any
objects on the second pass (the status log) which we imported
in the first pass (the objects.cache import).

Signed-off-by: Andreas Ericsson <ae@op5.se>