object_importer: Refactor setting object key
Previously we used to set the object id when we first located
it. This meant that we didn't know the name of the object, and
so couldn't look for a previously existing object with the same
name.

All in all, this felt a bit bug-prone, so we'd better change it
to determine the object id upon closing the object instead, which
seems to work much better in practice.

Signed-off-by: Andreas Ericsson <ae@op5.se>