object_importer: Be more careful when handling object indices
Since overwriting a previously existing object is a very bad
thing indeed, and we want to be able to spot such inconsistencies
right away, we add a helper class to take care of setting and
retrieving index values for the various objects.

This doesn't change the outcome of the import in any way, but
if it ever breaks again we'll know exactly why and where.

Signed-off-by: Andreas Ericsson <ae@op5.se>