specfile: Modify nagios.cfg in the module's %post section
Previously we used to do that when installing the daemon, which
also installed the database. However, the daemon can be run on
a system where no Nagios is running, so we must make sure not to
do that. The module, however, already requires monitor >= 4.0,
which in turn requires the correct configuration files. Thus, it
should be perfectly safe to modify nagios.cfg from the module's
%post section instead.

This fixes issue #2695.

Signed-off-by: Andreas Ericsson <ae@op5.se>