object_importer: Allow importing from multiple files
The check to verify that we weren't overwriting previously created
entries in \$this->imported wasn't careful enough, since it always
triggered when importing data from a second file. With a check to
see if the entry to be overwritten is identical to the one we're
about to overwrite it with (and let it pass if it is), importing
status data after having imported the object cache once again
works just fine.

Signed-off-by: Andreas Ericsson <ae@op5.se>