ipc.c: Use 0 msec timeout when poll()'ing for input
Since we'll only ever read from the socket when we already know
there's input waiting, and we'll only ever poll() it for *more*
input, we might as well poll() with a zero-millisecond timeout
to make sure we never waste time listening for something that
may not come.

This appears to decrease the time spent per event from a minimum
of 0.3 seconds (for a single event) down to less than 0.001
seconds, which is quite an astounding improvement.

Signed-off-by: Andreas Ericsson <ae@op5.se>