ipc: Handle socket read/write buffers intelligently
Previously we increased the buffer size of the passive socket
we used listen() for incoming connections on. While it might
be handy to have room for 128K simultaneous inbound connections,
it's far better to have all that space in the sockets that
actually do the work instead.

This patch makes it so, ensuring all accept()'ed or connect()'ed
sockets have their read/write buffers beefed up properly.

Signed-off-by: Andreas Ericsson <ae@op5.se>