daemon: fork() before importing objects and status
Running the import program in the background is important to prevent
letting the module flood our buffer with events, which would cause
us to re-run the import after a 15 second grace-time, which would
again cause the buffer to get flooded, and so on and so forth.

This patch makes the import program run in the background. It's
possible that we'll lose some events due to queries running before
the object they're targeting are present in the database, but that's
highly unlikely, and making sure we never run headlong into the
worst case scenario of hogging a CPU entirely by constantly running
the import program.

Signed-off-by: Andreas Ericsson <ae@op5.se>