gen-version: Include git commit id and '-dirty' in version names
Without the commit-id, we can't safely identify contributor's
patches from our own, so we need to have that there for when
git is built from the git repository (as all developers should
do).

This should also fix the case where people build from tarballs
and the version string goes missing.

Signed-off-by: Andreas Ericsson <ae@op5.se>