db_updater: Support NEBCALLBACK_CONTACT_NOTIFICATION events
The merlin daemon's sql code can now take care of inserting
NEBCALLBACK_CONTACT_NOTIFICATION events into the sql database.
We abuse the 'notification' table for this. The way to
distinguish between NEBCALLBACK_NOTIFICATION events and the
more fleshed out NEBCALLBACK_CONTACT_NOTIFICATION events is
to check if the 'contact_name' field is nul'ed out or not.
If it is, it's a notification summary (NEBCALLBACK_NOTIFICATION).
If it isn't, it's a NEBCALLBACK_CONTACT_NOTIFICATION event,
to be listed in the notification log view.

Signed-off-by: Andreas Ericsson <ae@op5.se>