ipc: Make ipc_is_connected() handle ipc_init() return code properly
Previously we used to just return whatever it handed us, but in
case the connection flaked out ipc_init() would return -1. A
careless evaluator might take that as a sign of connectedness
(since the function name has a boolean ring, that's not even
surprising). It appears we have such a caller and since we may
get more of them, we're better off taking care to make sure
ipc_is_connected() returns what's expected of it rather than
forcing each caller to look up exactly what it's supposed to
return in case we're not connected already.

This was part of the trigger for issue #2774.

Signed-off-by: Andreas Ericsson <ae@op5.se>