ipc: Use a merlin_event_counter to keep track of ipc events
This makes it possible to keep track of what happened to events
when and where and log it once in a bulk manner.

Each new session created resets the counter so the numbers should
always match up. This is an easy way of checking that the daemon
and the module is communicating without flaws.

Signed-off-by: Andreas Ericsson <ae@op5.se>