logging: Properly grok the log_level(s) variable
Previously we'd always bail out due to the fact that we had a NULL
sentinel in the array of valid opcodes but still looped it according
to ARRAY_SIZE().

Using a default log-level of "everything" but never resetting it
also meant we could never actually choose what we wanted to log.

This patch solves both those issues.

Signed-off-by: Andreas Ericsson <ae@op5.se>